- All the warnings in the 64-bit configuration were from implicit casting size_t values to ints.  I've made the casts explicit for now simply because if these values ever overflow a 32-bit integer, I'll have much bigger problems to worry about than truncation (like memory usage.  all of these size_t values come from container.size()). refs #111

+++ Source/WindowsApplication/Emunisce/InputRecording.cpp
@@ -70,7 +70,7 @@ void InputRecording::SerializeHistory(Archive& archive)
 {
 	SerializeItem(archive, m_recordingStartFrame);
 
-	unsigned int numEvents = m_inputHistory.size();
+	unsigned int numEvents = (unsigned int)m_inputHistory.size();
 	SerializeItem(archive, numEvents);
 
 	if(archive.GetArchiveMode() == ArchiveMode::Loading)
